Match-id-9abfa97de98d3df492229f01f58782796ec2e7f5

+++ src/framework/graph/core/node/legacy_node.cpp
@@ -127,7 +127,7 @@ LegacyNode::Vistor<OutDataAnchorPtr> LegacyNode::GetAllOutDataAnchors() const
 
 InDataAnchorPtr LegacyNode::GetInDataAnchor(int idx) const
 {
-    return ROLE(NodeStore).InDataAnchor(static_cast<std::size_t>(idx));
+    return ROLE(NodeStore).InDataAnchor(static_cast<unsigned int>(idx));
 }
 
 GraphErrCodeStatus LegacyNode::RemoveInDataAnchor(int idx)
@@ -143,7 +143,7 @@ GraphErrCodeStatus LegacyNode::RemoveInDataAnchor(int idx)
 
 OutDataAnchorPtr LegacyNode::GetOutDataAnchor(int idx) const
 {
-    return ROLE(NodeStore).OutDataAnchor(static_cast<std::size_t>(idx));
+    return ROLE(NodeStore).OutDataAnchor(static_cast<unsigned int>(idx));
 }
 
 InControlAnchorPtr LegacyNode::GetInControlAnchor() const
